Improve the SpecHelper patching code.
This splits it in to a module, adds support for @jkeiser's new DSL
management methods for 12.4, and makes sure to only patch in helpers
if a name hasn't been overridden by a nested scope.

This does break encapsulation on Chef::NodeMap because it doesn't
have an API to remove things, though this only matters in <=12.3 where
the NodeMap is shared by all resource classes. In 12.4, the NodeMap is
specific to each resource class and DescendantTracker acts as the
global registration system.