Remove the error on dependencies on prerelease's.
Now it trims to the non-alpha component and uses that. This is because
you have to depend on a prerelease Requirement to get it to accept a
prerelease that is before the desired target. For example, 2.0.pre does
not match ~>2.0. I'm not happy with this solution but it does unblock
me for now and I can keep pondering. Probably won't announce this
change.