Set the canonical resource name if possible.
This is important because in 12.4, it will see the parent class's .name
because self.inherited is processed before our .name is defined. This
means when subclassing something other than Chef::Resource in a test,
the canonical name was the parent. This will erase the incorrect
canonical name and set a new one.