converter: Write out gem dependencies to metadata.rb
This update writes out non-cookbook gem dependencies as "gem"
dependencies in the cookbook's metadata.rb file.

This should enhance the Halite development experience, ensuring that the
cookbook developer can accurately assume that their dependencies will be
pulled in, instead of having to mess around with chef_gem or what not.

The gem dependency logic pulls in runtime dependencies only and skips
the "halite" gem itself.